make events into a partial class like monomodrules
