+++ Celeste.Mod.mm/Patches/LevelEnter.cs
@@ -9,6 +9,9 @@ using MonoMod;
 using System;
 using System.Collections;
 
+using _LevelEnter = Celeste.LevelEnter;
+using _Session = Celeste.Session;
+
 namespace Celeste {
     // LevelEnter has a private .ctor
     class patch_LevelEnter : Scene {
@@ -184,5 +187,41 @@ namespace Celeste {
             }
         }
 
+    }
+}
+
+namespace Celeste.Mod {
+    public static partial class Everest {
+        public static partial class Events {
+            public static partial class Level {
+                public delegate void EnterHandler(_Session session, bool fromSaveData);
+                /// <summary>
+                /// Called at the end of <see cref="patch_LevelEnter.Go"/>, provided no error occurred.
+                /// </summary>
+                public static event EnterHandler OnEnter;
+                internal static void Enter(_Session session, bool fromSaveData)
+                    => OnEnter?.Invoke(session, fromSaveData);
+            }
+            public static class LevelEnter {
+                public delegate Scene CustomVignetteHandler(_Session session, bool fromSaveData);
+                /// <summary>
+                /// Called at the beginning of <see cref="_LevelEnter.PlayCustomVignette()"/>.<br/>
+                /// Invoked when entering a map, before instantiating and setting the next scene.<br/>
+                /// Should return the next scene to set if some conditions are met, or null otherwise.
+                /// </summary>
+                public static event CustomVignetteHandler OnPlayVignette;
+                internal static Scene PlayVignette(_Session session, bool fromSaveData) {
+                    if (OnPlayVignette is null)
+                        return null;
+
+                    foreach (CustomVignetteHandler handler in OnPlayVignette.GetInvocationList()) {
+                        if (handler(session, fromSaveData) is { } nextScene)
+                            return nextScene;
+                    }
+
+                    return null;
+                }
+            }
+        }
     }
 }
