+++ Celeste.Mod.mm/Patches/OuiJournal.cs
@@ -1,7 +1,10 @@
 #pragma warning disable CS0626 // Method, operator, or accessor is marked external and has no attributes on it
 
 using Celeste.Mod;
+using System;
 using System.Collections;
+
+using _OuiJournal = Celeste.OuiJournal;
 
 namespace Celeste {
     class patch_OuiJournal : OuiJournal {
@@ -30,5 +33,35 @@ namespace Celeste {
             }
         }
 
+    }
+}
+
+namespace Celeste.Mod {
+    public static partial class Everest {
+        public static partial class Events {
+            [Obsolete("Use Journal instead.")]
+            public static class OuiJournal {
+                public delegate void EnterHandler(_OuiJournal journal, Oui from);
+                public static event EnterHandler OnCreateButtons {
+                    add {
+                        Journal.OnEnter += (Journal.EnterHandler) value.CastDelegate(typeof(Journal.EnterHandler));
+                    }
+                    remove {
+                        Journal.OnEnter -= (Journal.EnterHandler) value.CastDelegate(typeof(Journal.EnterHandler));
+                    }
+                }
+            }
+
+            public static class Journal {
+                public delegate void EnterHandler(_OuiJournal journal, Oui from);
+                /// <summary>
+                /// Called by <see cref="patch_OuiJournal.Enter"/>.
+                /// </summary>
+                public static event EnterHandler OnEnter;
+                internal static void Enter(_OuiJournal journal, Oui from)
+                    => OnEnter?.Invoke(journal, from);
+            }
+
+        }
     }
 }
